services/timeseries-service/requirements.txt
kafka-python==2.0.2
psycopg2-binary==2.9.9
minio==7.1.17
python-dotenv==1.0.0
fastapi==0.104.0
uvicorn==0.23.2
